Moved training plan analyzers to folder. Moved logic around. Minor changes
